# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Teemer Moving & Storage API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customer-auth/signup:
    post:
      operationId: customerSignup
      tags: [customer-auth]
      summary: Create a customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerSignupRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSignupResponse"
        "400":
          description: Invalid input
        "409":
          description: Email or username already in use
  /customer-auth/login:
    post:
      operationId: customerLogin
      tags: [customer-auth]
      summary: Customer login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerLoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSessionResponse"
        "401":
          description: Invalid credentials
...
  /customer-auth/me:
    get:
      operationId: customerMe
      tags: [customer-auth]
      summary: Current signed-in customer
      responses:
        "200":
          description: Current customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSessionResponse"
        "401":
          description: Not authenticated
  /customer-auth/check-username:
    get:
      operationId: customerCheckUsername
      tags: [customer-auth]
      summary: Check if a username is available
      parameters:
        - in: query
          name: username
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Availability info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameAvailability"
...
  /admin/customers/lookup:
    get:
      operationId: lookupCustomers
      tags: [admin]
      summary: Look up customers by name, email, phone, or username
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Matching customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminCustomerLookup"
  /admin/payment-requests:
    get:
      operationId: listAdminPaymentRequests
      tags: [admin]
      summary: List all payment requests
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: Payment requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminPaymentRequest"
    post:
      operationId: createAdminPaymentRequest
      tags: [admin]
      summary: Send a payment request to a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequestRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPaymentRequest"
  /admin/payments:
    get:
      operationId: listAdminPayments
      tags: [admin]
      summary: List all payments (admin Payments view)
      parameters:
        - in: query
          name: method
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminPaymentRow"
components:
  schemas:
...
    CustomerSignupRequest:
      type: object
      required: [fullName, email, phone]
      properties:
        fullName: { type: string }
        email: { type: string }
        phone: { type: string }
        username:
          type: string
          description: "Optional. If omitted, generated from name. Format: +letters/digits/underscores 3-30 chars."
        password:
          type: string
          description: Optional. If omitted a strong password is generated and emailed.
        attachQuoteId:
          type: integer
          nullable: true
    CustomerSignupResponse:
      type: object
      properties:
        customer:
          $ref: "#/components/schemas/CustomerSession"
        generatedPassword:
          type: string
          nullable: true
          description: The auto-generated password if no password was supplied. Shown once.
    CustomerLoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Username (with leading +) or email
        password: { type: string }
    CustomerSession:
      type: object
      required: [customerId, email, username, fullName]
      properties:
        customerId: { type: integer }
        email: { type: string }
        phone: { type: string, nullable: true }
        username: { type: string }
        fullName: { type: string }
    CustomerSessionResponse:
      type: object
      properties:
        customer:
          $ref: "#/components/schemas/CustomerSession"
    UsernameAvailability:
      type: object
      properties:
        available: { type: boolean }
        valid: { type: boolean }
        normalized: { type: string }
...
    AdminCustomerLookup:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string }
        email: { type: string }
        phone: { type: string }
        username: { type: string, nullable: true }
        hasAccount: { type: boolean }
    CreatePaymentRequestRequest:
      type: object
      required: [amountCents, description]
      properties:
        username: { type: string, nullable: true }
        customerId: { type: integer, nullable: true }
        amountCents: { type: integer, minimum: 100 }
        description: { type: string }
    AdminPaymentRequest:
      type: object
      properties:
        id: { type: integer }
        customerId: { type: integer }
        amountCents: { type: integer }
        description: { type: string }
        status: { type: string }
        stripeSessionId: { type: string, nullable: true }
        confirmationNumber: { type: string, nullable: true }
        createdAt: { type: string, nullable: true }
        paidAt: { type: string, nullable: true }
        payUrl: { type: string, nullable: true }
        customer:
          type: object
          nullable: true
          properties:
            id: { type: integer }
            fullName: { type: string }
            email: { type: string }
            username: { type: string, nullable: true }
    AdminPaymentRow:
      type: object
      properties:
        id: { type: integer }
        jobId: { type: integer, nullable: true }
        customerId: { type: integer, nullable: true }
        paymentRequestId: { type: integer, nullable: true }
        type: { type: string }
        method: { type: string, nullable: true }
        amount: { type: number }
        reference: { type: string, nullable: true }
        confirmationNumber: { type: string, nullable: true }
        paidAt: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        job:
          type: object
          nullable: true
        quote:
          type: object
          nullable: true
        customer:
          type: object
          nullable: true