# lib/api-zod/src/generated/types/adminPaymentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teemer Moving & Storage API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminPaymentRequestCustomer } from "./adminPaymentRequestCustomer";

export interface AdminPaymentRequest {
  id?: number;
  customerId?: number;
  amountCents?: number;
  description?: string;
  status?: string;
  stripeSessionId?: string | null;
  confirmationNumber?: string | null;
  createdAt?: string | null;
  paidAt?: string | null;
  payUrl?: string | null;
  customer?: AdminPaymentRequestCustomer;
}