# lib/api-zod/src/generated/types/adminPaymentRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teemer Moving & Storage API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminPaymentRowCustomer } from "./adminPaymentRowCustomer";
import type { AdminPaymentRowJob } from "./adminPaymentRowJob";
import type { AdminPaymentRowQuote } from "./adminPaymentRowQuote";

export interface AdminPaymentRow {
  id?: number;
  jobId?: number | null;
  customerId?: number | null;
  paymentRequestId?: number | null;
  type?: string;
  method?: string | null;
  amount?: number;
  reference?: string | null;
  confirmationNumber?: string | null;
  paidAt?: string | null;
  notes?: string | null;
  job?: AdminPaymentRowJob;
  quote?: AdminPaymentRowQuote;
  customer?: AdminPaymentRowCustomer;
}