# lib/api-zod/src/generated/types/customerLoginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teemer Moving & Storage API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerLoginRequest {
  /** Username (with leading +) or email */
  identifier: string;
  password: string;
}